Implement revision suggestions. As test is a sandbox remove large files and add readme and test notes to locate them before running the test.
