+++ hapi-fhir-jpaserver-test-utilities/src/test/resources/loinc-full/readme.txt
@@ -1,0 +1,10 @@
+LoincFullLoadR4SandboxIT requires this directory must contain the following
+three files:
+   _ Loinc.csv.gz
+   _ Loinc_1.11.zip and
+   _ v1.11_loincupload.properties
+
+but those files are too large for the repo, so before running this test you need to copy them from:
+https://drive.google.com/drive/folders/18be2R5IurlWnugkl18nDG7wrwPsOtfR-?usp=sharing
+(SmileCDR has access)
+
