--- hapi-fhir-jpaserver-test-utilities/src/test/resources/loinc-full/v1.11_loincupload.properties
@@ -1,89 +1,0 @@
-#################
-### MANDATORY ###
-#################
-
-# Answer lists (ValueSets of potential answers/values for LOINC "questions")
-## File must be present
-loinc.answerlist.file=AccessoryFiles/AnswerFile/AnswerList.csv
-# Answer list links (connects LOINC observation codes to answer list codes)
-## File must be present
-loinc.answerlist.link.file=AccessoryFiles/AnswerFile/LoincAnswerListLink.csv
-
-# Document ontology
-## File must be present
-loinc.document.ontology.file=AccessoryFiles/DocumentOntology/DocumentOntology.csv
-
-# LOINC codes
-## File must be present
-loinc.file=LoincTable/Loinc.csv
-
-# LOINC hierarchy
-## File must be present
-loinc.hierarchy.file=AccessoryFiles/ComponentHierarchyBySystem/ComponentHierarchyBySystem.csv
-
-# IEEE medical device codes
-## File must be present
-loinc.ieee.medical.device.code.mapping.table.file=AccessoryFiles/LoincIeeeMedicalDeviceCodeMappingTable/LoincIeeeMedicalDeviceCodeMappingTable.csv
-
-# Imaging document codes
-## File must be present
-loinc.imaging.document.codes.file=AccessoryFiles/ImagingDocuments/ImagingDocumentCodes.csv
-
-# Part
-## File must be present
-loinc.part.file=AccessoryFiles/PartFile/Part.csv
-
-# Part link
-## File must be present
-loinc.part.link.primary.file=AccessoryFiles/PartFile/LoincPartLink_Primary.csv
-loinc.part.link.supplementary.file=AccessoryFiles/PartFile/LoincPartLink_Supplementary.csv
-
-# Part related code mapping
-## File must be present
-loinc.part.related.code.mapping.file=AccessoryFiles/PartFile/PartRelatedCodeMapping.csv
-
-# RSNA playbook
-## File must be present
-loinc.rsna.playbook.file=AccessoryFiles/LoincRsnaRadiologyPlaybook/LoincRsnaRadiologyPlaybook.csv
-
-# Top 2000 codes - SI
-## File must be present
-loinc.top2000.common.lab.results.si.file=AccessoryFiles/Top2000Results/SI/Top2000CommonLabResultsSi.csv
-# Top 2000 codes - US
-## File must be present
-loinc.top2000.common.lab.results.us.file=AccessoryFiles/Top2000Results/US/Top2000CommonLabResultsUs.csv
-
-# Universal lab order ValueSet
-## File must be present
-loinc.universal.lab.order.valueset.file=AccessoryFiles/LoincUniversalLabOrdersValueSet/LoincUniversalLabOrdersValueSet.csv
-
-################
-### OPTIONAL ###
-################
-
-# This is the version identifier for the LOINC code system
-## Key may be omitted if only a single version of LOINC is being kept.
-loinc.codesystem.version=1.11
-
-loinc.codesystem.make.current=false
-
-# This is the version identifier for the answer list file
-## Key may be omitted
-loinc.answerlist.version=Beta.1
-
-# This is the version identifier for uploaded ConceptMap resources
-## Key may be omitted
-loinc.conceptmap.version=Beta.1
-
-# Group
-## Default value if key not provided: AccessoryFiles/GroupFile/Group.csv
-## File may be omitted
-loinc.group.file=AccessoryFiles/GroupFile/Group.csv
-# Group terms
-## Default value if key not provided: AccessoryFiles/GroupFile/GroupLoincTerms.csv
-## File may be omitted
-loinc.group.terms.file=AccessoryFiles/GroupFile/GroupLoincTerms.csv
-# Parent group
-## Default value if key not provided: AccessoryFiles/GroupFile/ParentGroup.csv
-## File may be omitted
-loinc.parent.group.file=AccessoryFiles/GroupFile/ParentGroup.csv
